gym-ops 1.1

+++ Back/src/modules/exercicios/service/ExercicioService.ts
@@ -1,17 +1,21 @@
 import { inject, injectable } from "tsyringe";
-import { ExercicioRepository } from "../repository/ExercicioRepository";
 import { ExercicioSchema } from "../dto/ExercicioSchema";
 import { z } from "zod";
 import AppError from "../../../shared/errors/AppError";
-import UserRepository from "../../user/repositories/UserRepository";
+import { getPaginationOffset } from "../../../shared/helpers/getPaginationOffset";
+import { IExercicioRepository } from "../interface/IExercicioRepository";
+import { IUserRepository } from "../../user/interface/IUserRepository";
+import { IExercicioDeTreinoRepository } from "../../exerciciosDeTreinos/interface/IExercicioDeTreinoRepository";
 
 @injectable()
 export class ExercicioService {
   constructor(
-    @inject(ExercicioRepository)
-    private exercicioRepository: ExercicioRepository,
-    @inject(UserRepository)
-    private userRepository: UserRepository,
+    @inject("ExercicioRepository")
+    private exercicioRepository: IExercicioRepository,
+    @inject("UserRepository")
+    private userRepository: IUserRepository,
+    @inject("ExercicioDeTreinoRepository")
+    private exercicioDeTreinoRepository: IExercicioDeTreinoRepository,
   ) {}
 
   async create(data: z.infer<typeof ExercicioSchema>, adm_id: number) {
@@ -34,8 +38,9 @@ export class ExercicioService {
     return await this.exercicioRepository.create(exercicioData);
   }
 
-  async list(adm_id: number) {
-    return this.exercicioRepository.list(adm_id);
+  async list(adm_id: number, limit: number, page: number) {
+    const offset = getPaginationOffset(page, limit);
+    return this.exercicioRepository.list(adm_id, offset, limit);
   }
 
   async update(
@@ -69,6 +74,12 @@ export class ExercicioService {
       throw new AppError("Exercício não encontrado", 404);
     }
 
+    const InWorkouts =
+      await this.exercicioDeTreinoRepository.exerciseInWorkouts(id, adm_id);
+
+    if (InWorkouts) {
+      throw new AppError("Exercício está em uso em um treino", 409);
+    }
     return await this.exercicioRepository.delete(id, adm_id);
   }
